feat(callbacks):启用回调逻辑 01.优化获取寄存器部分的代码 02.补充代码注释、调整变量命名

+++ main.c
@@ -4,6 +4,7 @@
 int globalexit = 0;
 int printMsg(struct CbMsg *info)
 {
+    if(!info) return -1;
     enum MsgLevel level = ML_INFO;
     switch (info->ocb) {
     case ID_CLOSE:
@@ -17,27 +18,27 @@ int printMsg(struct CbMsg *info)
     }
 
     do{
-        if(info->path && (info->path[0] == 'p' || info->path[0] == 's'))
+        if(info->opath && (info->opath[0] == 'p' || info->opath[0] == 's'))
             break;
 //        DMSG(level,
 //             "info.otid = %llu\tinfo.type = %llu\t"
 //             "info.gpid = %llu\tinfo.pid = %llu\n",
 //             info->otid,info->type,info->gpid,info->pid);
-        if(info->exe)
-            DMSG(level,
-                 "info.exelen  = %llu\tinfo.exe = %s\n",
-                 info->exelen,info->exe);
-        if(info->path)
-        {
-            if(info->path[0] != '/') break;
-            DMSG(level,
-                 "info.pathlen = %llu\tinfo.path = %s\n",
-                 info->pathlen,info->path);
-        }
+        if(info->opath[0] != '/') break;
+        if(!info->tpath)
+            DMSG(level,"Target exe %s operates on file %s\n",
+                 info->exe,info->opath);
+        else
+            DMSG(level,"Target exe %s operates from file %s to %s\n",
+                 info->exe,info->opath,info->tpath);
     }while(0);
 
-    if(info->exe)  free(info->exe);
-    if(info->path) free(info->path);
+//    if(info->exe)
+//    {
+//        free(info->exe);
+//        info->exe = NULL;
+//    }
+    if(info->opath) free(info->opath);
     if(info) free(info);
     return 0;
 }
@@ -46,7 +47,7 @@ int printMsg(struct CbMsg *info)
 //extern InitInfo gInitInfo[PROC_MAX];
 static void sigOptions(int sig)
 {
-    printf(" signal = %d\n",sig);
+    printf(" signal %d\n",sig);
     if(sig == SIGINT || sig == SIGTERM)
     {
         DMSG(ML_WARN,"\n正在通知退出，请稍候...\n");
@@ -62,23 +63,24 @@ static void sigOptions(int sig)
     }
 }
 
+// 标准输出重定向
+void redirectStdout()
+{
+    char *filename = "output.log";
+    FILE *fp = freopen(filename,"w+",stdout);       //标准输出重定向到文件output.log
+    if(fp)
+        DMSG(ML_INFO,"Redirect stdout to %s\n", filename);
+    else
+        DERR(freopen);
+}
+
 int main(int argc, char** argv)
 {
-//    FILE *fp = NULL;
-//    char *filename = "output.log";
-
-//    fp = freopen(filename,"w+",stdout);       //标准输出重定向到文件output.log
-//    if (NULL == fp)
-//    {
-//        fprintf(stderr,"error redirecting stdout\n");
-//        return 0;
-//    }
-//    printf("Redirect stdout to %s\n", filename);
-
-    // 设置进程优先级 -20 是最高优先级
+//    redirectStdout();
+    // 设置进程优先级 -20 是最高优先级(子进程会继承该优先级)
     setpriority(PRIO_PROCESS, getpid(), -20);
     // 启动监控进程
-    StartSystemMonitor();
+    if(StartSystemMonitor()) return -1;
     // 给主进程单独设置信号处理函数
     signal(SIGINT,sigOptions);  // Ctrl + c
     signal(SIGTERM,sigOptions); // kill -15
